python/requirements.txt
fluent-logger==0.9.6
msgpack==0.6.2
